roachpb: move static methods outside of generated code
Also avoids using plain strings in panic objects.

Release note: None
