+++ pkg/roachpb/batch.go
@@ -14,6 +14,7 @@ import (
 	"bytes"
 	"context"
 	"fmt"
+	"strconv"
 	"strings"
 
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/concurrency/lock"
@@ -36,6 +37,37 @@ func (h Header) WriteTimestamp() hlc.Timestamp {
 		ts.Forward(h.Txn.WriteTimestamp)
 	}
 	return ts
+}
+
+// Summary prints a short summary of the requests in a batch.
+func (ba *BatchRequest) Summary() string {
+	var b strings.Builder
+	ba.WriteSummary(&b)
+	return b.String()
+}
+
+// WriteSummary writes a short summary of the requests in a batch
+// to the provided builder.
+func (ba *BatchRequest) WriteSummary(b *strings.Builder) {
+	if len(ba.Requests) == 0 {
+		b.WriteString("empty batch")
+		return
+	}
+	counts := ba.getReqCounts()
+	var tmp [10]byte
+	var comma bool
+	for i, v := range counts {
+		if v != 0 {
+			if comma {
+				b.WriteString(", ")
+			}
+			comma = true
+
+			b.Write(strconv.AppendInt(tmp[:0], int64(v), 10))
+			b.WriteString(" ")
+			b.WriteString(requestNames[i])
+		}
+	}
 }
 
 // SetActiveTimestamp sets the correct timestamp at which the request is to be
